# ms-gateaway/docs/swagger.yaml
basePath: /
definitions:
  auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  auth.RegisterRequest:
    properties:
      address:
        type: string
      email:
        type: string
      full_name:
        type: string
      password:
        type: string
    type: object
  auth.RegisterResponse:
    properties:
      address:
        type: string
      balance:
        type: number
      email:
        type: string
      full_name:
        type: string
      role:
        type: string
      user_id:
        type: integer
    type: object
  helper.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is an auth service.
  title: Auth Service API
  version: "1.0"
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: Login a user with email and password
      parameters:
      - description: Login Request
        in: body
        name: loginRequest
        required: true
        schema:
          $ref: '#/definitions/auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.HTTPError'
      summary: Login a user
      tags:
      - auth
...
  /register:
    post:
      consumes:
      - application/json
      description: Register a new user with full details
      parameters:
      - description: Register Request
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/auth.RegisterResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/helper.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/helper.HTTPError'
      summary: Register a new user
      tags:
      - auth
...
swagger: "2.0"